requirements.txt
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.1
matplotlib>=3.8.0
seaborn>=0.13.2
streamlit>=1.42.0
plotly>=5.24.0
